requirements.txt
mcp
openai
httpx
uvicorn
python-dotenv
